requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
rich>=13.0.0
python-nmap>=0.7.1
concurrent-log-handler>=0.9.24